Invite users on Cody Dashboard
