+++ client/web/BUILD.bazel
@@ -231,6 +231,7 @@ ts_project(
         "src/cody/dashboard/UpsellImage.tsx",
         "src/cody/editorGroups.ts",
         "src/cody/invites/AcceptInviteBanner.tsx",
+        "src/cody/invites/InviteUsers.tsx",
         "src/cody/invites/useInviteParams.ts",
         "src/cody/invites/useInviteState.ts",
         "src/cody/isCodyEnabled.tsx",
@@ -282,7 +283,6 @@ ts_project(
         "src/cody/subscription/queries.tsx",
         "src/cody/switch-account/CodySwitchAccountPage.tsx",
         "src/cody/team/CodyManageTeamPage.tsx",
-        "src/cody/team/InviteUsers.tsx",
         "src/cody/team/TeamMemberList.tsx",
         "src/cody/upsell/ChatBrandIcon.tsx",
         "src/cody/upsell/CodyUpsellPage.tsx",
